Multi Agent RAG/backend/requirements.txt
fastapi
pymongo
uvicorn[standard]
python-dotenv

langchain
langgraph
langchain-core
langchain-community

httpx
duckduckgo-search